Python 3 comptability, use of collections.abc
