+++ .travis.yml
@@ -1,7 +1,18 @@
+sudo: false
 language: python
 python:
   - "2.7"
-  - "2.6"
+  - "3.3"
+  - "3.4"
+  - "3.5"
+  - "3.6"
+  - "nightly"
   - "pypy"
+  - "pypy3"
 install: true
+matrix:
+  fast_finish: true
+  allow_failures:
+  - python: "3.6"
+  - python: "nightly"
 script:  nosetests
